refactor: move out run command
